devimint

Module federation

Modules

config 🔒

Structs

Client

fedimint-cli instance (basically path with client state: config + db)

Federation

Fedimintd

Constants

FEDIMINTD_API_PORT_OFFSET

Which port is for api inside the range from PORTS_PER_FEDIMINTD

FEDIMINTD_METRICS_PORT_OFFSET

Which port is for prometheus inside the range from PORTS_PER_FEDIMINTD

FEDIMINTD_P2P_PORT_OFFSET

Which port is for p2p inside the range from PORTS_PER_FEDIMINTD

FEDIMINTD_UI_PORT_OFFSET

Which port is for the web ui inside the range from PORTS_PER_FEDIMINTD

PORTS_PER_FEDIMINTD

Number of ports we allocate for every fedimintd instance

Functions

cli_set_config_gen_params 🔒

cli_wait_server_status 🔒

run_cli_dkg

run_cli_dkg_v2