devimint

Module run_devfed_test_builder

Tools for manipulating the type state of RunDevfedTestBuilder.

See the detailed guide that describes how all the pieces here fit together.

Modules

sealed 🔒

Structs

Empty

Represents a State that has IsUnset implemented for all members.

SetNumFeds

Represents a State that has IsSet implemented for State::NumFeds.

SetPreDkg

Represents a State that has IsSet implemented for State::PreDkg.

Traits

IsComplete

Marker trait that indicates that all required members are set.

IsSet

Marker trait that indicates that the member is set, i.e. at least one of its setters was called.

IsUnset

Marker trait that indicates that the member is unset, i.e. none of its setters was called.

State

Builder’s type state specifies if members are set or not (unset).