devimint::run_devfed_test_builder

Trait IsComplete

pub trait IsComplete: State { }

Marker trait that indicates that all required members are set.

In this state, you can finish building by calling the method RunDevfedTestBuilder::call()

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.