devimint

Module tests

Structs

Stats

UpgradeClients

Clients reused for upgrade tests

Enums

LatencyTest

TestCmd

UpgradeTest

Functions

all_peer_block_count ğŸ”’

cannot_replay_tx_test

cli_load_test_tool_test

cli_tests

do_try_create_and_pay_invoice

guardian_backup_test

gw_reboot_test

handle_command

latency_tests

lightning_gw_reconnect_test

ln_invoice ğŸ”’

ln_pay ğŸ”’

lnv2_receive ğŸ”’

lnv2_send ğŸ”’

log_binary_versions

modify_server_configs ğŸ”’

Modify server configuration files to add a new meta module instance

modify_single_peer_config ğŸ”’

Modify configuration files for a single peer to add a new meta module instance

peer_block_count ğŸ”’

reconnect_test

recoverytool_test

run_ln_circular_load_test

run_standard_load_test

stats_for

stress_test_fed ğŸ”’

test_client_config_change_detection

Test that client can detect federation config changes when servers restart with new module configurations

test_guardian_password_change

test_offline_client_initialization

Test that client can init even when the federation is down

upgrade_tests