devimint

Module util

Modules

legacy_types 🔒

Structs

BitcoinCli

Bitcoind

Command

DevimintFaucet

Esplora

FedimintCli

FedimintdCmd

GatewayCli

GatewayLdkCli

GatewayLndCli

Gatewayd

LnCli

Lnd

LoadTestTool

ProcessHandle

Kills process when all references to ProcessHandle are dropped.

ProcessHandleInner

ProcessManager

Recoverytool

Recurringd

Constants

BITCOIND_FALLBACK 🔒

BITCOIN_CLI_FALLBACK 🔒

DEFAULT_POLL_TIMEOUT 🔒

DEFAULT_VERSION 🔒

DEVIMINT_FAUCET_FALLBACK 🔒

ESPLORA_FALLBACK 🔒

FEDIMINTD_FALLBACK 🔒

FEDIMINT_CLI_FALLBACK 🔒

FEDIMINT_DBTOOL_FALLBACK 🔒

GATEWAYD_FALLBACK 🔒

GATEWAY_CLI_FALLBACK 🔒

LNCLI_FALLBACK 🔒

LND_FALLBACK 🔒

LOAD_TEST_TOOL_FALLBACK 🔒

RECOVERYTOOL_FALLBACK 🔒

Traits

JsonValueExt

ToCmdExt

Functions

get_bitcoin_cli_path

get_command_str_for_alias 🔒

get_fedimint_cli_path

get_fedimint_dbtool_cli_path

get_gateway_cli_path

get_lncli_path

is_backwards_compatibility_test

Returns true if running backwards-compatibility tests

parse_clap_version 🔒

Parses a version string returned from clap ex: fedimintd 0.3.0-alpha -> 0.3.0-alpha

parse_map

poll

Retry until f succeeds or default timeout is reached

poll_simple

poll_with_timeout

Retry until f succeeds or timeout is reached

send_sigkill 🔒

send_signal 🔒

send_sigterm 🔒

supports_lnv2

to_command 🔒

use_fedimint_cli

Sets the fedimint-cli and mint client alias

use_matching_fedimint_cli_for_dkg

Sets the fedimint-cli binary to match the fedimintd’s version, which is needed for running DKG. Returns the original fedimint-cli path and mint client alias so the caller can reset the fedimint-cli version after DKG

version_hash_to_version 🔒

Maps a version hash to a release version