devimint
Module envs

Constants

FM_BACKWARDS_COMPATIBILITY_TEST_ENV

FM_BITCOIND_BASE_EXECUTABLE_ENV

FM_BITCOIN_CLI_BASE_EXECUTABLE_ENV

FM_BITCOIN_RPC_URL_ENV

FM_BTC_CLIENT_ENV

FM_CLIENT_DIR_ENV

FM_CLN_SOCKET_ENV

FM_DATA_DIR_ENV

FM_DEVIMINT_CMD_INHERIT_STDERR_ENV

Make devimint print stderr of called commands directly on its own stderr

FM_DEVIMINT_RUN_DEPRECATED_TESTS_ENV

Force devimint to run a test with a deprecated configuration

FM_ELECTRS_BASE_EXECUTABLE_ENV

FM_ESPLORA_BASE_EXECUTABLE_ENV

FM_FAUCET_BASE_EXECUTABLE_ENV

FM_FAUCET_BIND_ADDR_ENV

FM_FEDIMINTD_BASE_EXECUTABLE_ENV

FM_FEDIMINT_CLI_BASE_EXECUTABLE_ENV

FM_FEDIMINT_DBTOOL_BASE_EXECUTABLE_ENV

FM_FED_SIZE_ENV

FM_GATEWAYD_BASE_EXECUTABLE_ENV

FM_GATEWAY_API_ADDR_ENV

FM_GATEWAY_CLI_BASE_EXECUTABLE_ENV

FM_GATEWAY_CLN_EXTENSION_BASE_EXECUTABLE_ENV

FM_GATEWAY_DATA_DIR_ENV

FM_GATEWAY_LISTEN_ADDR_ENV

FM_GWCLI_CLN_ENV

FM_GWCLI_LND_ENV

FM_INVITE_CODE_ENV

FM_LIGHTNINGD_BASE_EXECUTABLE_ENV

FM_LIGHTNING_CLI_BASE_EXECUTABLE_ENV

FM_LIGHTNING_CLI_ENV

FM_LINK_TEST_DIR_ENV

FM_LNCLI_BASE_EXECUTABLE_ENV

FM_LNCLI_ENV

FM_LND_BASE_EXECUTABLE_ENV

FM_LOAD_TEST_TOOL_BASE_EXECUTABLE_ENV

FM_LOGS_DIR_ENV

FM_MINT_CLIENT_ENV

FM_OFFLINE_NODES_ENV

FM_PASSWORD_ENV

FM_PORT_GW_LND_ENV

FM_RECOVERYTOOL_BASE_EXECUTABLE_ENV

FM_TEST_DIR_ENV