Module devimint::util

Structs

Command

ProcessHandle

Kills process when all references to ProcessHandle are dropped.

ProcessHandleInner

ProcessManager

Constants

POLL_INTERVAL 🔒

Traits

JsonValueExt

ToCmdExt

Functions

poll

poll_max_retries

Will retry calling f until it returns Ok(true) or retries times. A notable difference from poll is that f may fail with an error at any time and we will still keep retrying.

poll_value

send_sigkill 🔒

send_signal 🔒

send_sigterm 🔒